+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -51,7 +51,7 @@ public class DisabledHub : IHub, IDisposable
     /// <summary>
     /// Returns a dummy transaction.
     /// </summary>
-    public ITransaction StartTransaction(
+    public ITransactionTracer StartTransaction(
         ITransactionContext context,
         IReadOnlyDictionary<string, object?> customSamplingContext) =>
         // Transactions from DisabledHub are always sampled out
@@ -60,14 +60,14 @@ public class DisabledHub : IHub, IDisposable
     /// <summary>
     /// No-Op.
     /// </summary>
-    public void BindException(Exception exception, ISpan span)
+    public void BindException(Exception exception, ISpanTracer span)
     {
     }
 
     /// <summary>
     /// Returns null.
     /// </summary>
-    public ISpan? GetSpan() => null;
+    public ISpanTracer? GetSpan() => null;
 
     /// <summary>
     /// Returns null.