+++ CHANGELOG.md
@@ -33,6 +33,7 @@ API Changes:
 - The `DiagnosticLogger` signature for `LogError` and `LogFatal` changed to take the `exception` as the first parameter. That way it does no longer get mixed up with the TArgs. The `DiagnosticLogger` now also received an overload for `LogError` and `LogFatal` that accepts a message only. ([#2715](https://github.com/getsentry/sentry-dotnet/pull/2715))
 - The `Extra` property and `SetExtra` method on `Span`, `Scope` and `Transaction` have been renamed to `Data` and `SetData` for consistency with Sentry's other SDKs. You'll need to update any references to these properties/methods in your own code. ([#2730](https://github.com/getsentry/sentry-dotnet/pull/2730))
 - Integrate `sentry-native` as a static library in Native AOT builds to enable symbolication. ([2704](https://github.com/getsentry/sentry-dotnet/pull/2704))
+- `ISpan` and `ITransaction` have been renamed to `ISpanTracer` and `ITransactionTracer`. You will need to update any references to these interfaces in your code to use the new interface names ([#2731](https://github.com/getsentry/sentry-dotnet/pull/2731))
 
 ## Unreleased